restructured project + changed to cmake
