+++ CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.18.4)
+
+project(simpleRenderer LANGUAGES CXX)
+
+add_subdirectory(src)
