+++ src/input/Input.cpp
@@ -1,4 +1,4 @@
-#include <Input.hpp>
+#include "Input.hpp"
 
 #include <GLFW/glfw3.h>
