+++ src/util/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(util
+    Util.hpp
+    Util.cpp
+)
+
+target_include_directories(util PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
